Updated classes

+++ src/DataMapper.php
@@ -60,6 +60,9 @@ class DataMapper
     /** @var bool  */
     protected $deleted = false;
 
+    /** @var array */
+    protected $pendingLinks = [];
+
     /**
      * DataMapper constructor.
      * @param EntityManager $entityManager
@@ -299,13 +302,22 @@ class DataMapper
     public function link(string $relation, $items)
     {
         $relations = $this->mapper->getRelations();
-        if(!isset($relation[$relation])){
+        if(!isset($relations[$relation])){
             throw new RuntimeException("Unknown relation '$relation'");
         }
 
         /** @var $rel HasOneOrManyThrough */
         if(!(($rel = $relations[$relation]) instanceof HasOneOrManyThrough)){
             throw new RuntimeException("Unsupported relation type");
+        }
+
+        if($this->isNew){
+            $this->pendingLinks[] = [
+                'relation' => $rel,
+                'items' => $items,
+                'link' => true,
+            ];
+            return;
         }
 
         $rel->link($this, $items);
@@ -325,6 +337,15 @@ class DataMapper
         /** @var $rel HasOneOrManyThrough */
         if(!(($rel = $relations[$relation]) instanceof HasOneOrManyThrough)){
             throw new RuntimeException("Unsupported relation type");
+        }
+
+        if($this->isNew){
+            $this->pendingLinks[] = [
+                'relation' => $rel,
+                'items' => $items,
+                'link' => false,
+            ];
+            return;
         }
 
         $rel->unlink($this, $items);
@@ -411,7 +432,8 @@ class DataMapper
             case 'string':
                 return (string) $value;
             case 'date':
-                return $value;
+                /** @var $value DateTime */
+                return $value->format($this->manager->getDateFormat());
             case 'json':
             case 'json-assoc':
                 return json_encode($value);
@@ -458,5 +480,20 @@ class DataMapper
         };
 
         return $closure->call($relation, $this, $callback);
+    }
+
+    protected function executePendingLinkage()
+    {
+        foreach ($this->pendingLinks as $item){
+            /** @var HasOneOrManyThrough $rel */
+            $rel = $item['relation'];
+            if($item['link']){
+                $rel->link($this, $item['items']);
+            } else {
+                $rel->unlink($this, $item['items']);
+            }
+        }
+
+        $this->pendingLinks = [];
     }
 }
